test: move test files into standard layout
Move the tests and test data into the standard Apache layout [1]. Using a
standard layout helps modernize the developer experience in the project and
will make the introduction of Maven as the preferred build system in a later
patch easier.

[1] https://maven.apache.org/guides/introduction/introduction-to-the-standard-directory-layout.html
