+++ src/test/resources/XMLParser/1-empty-project.pnd
