+++ src/test/resources/XMLParser/11-two-moves-one-rank-song-constants-renamed-command.pnd
