+++ src/test/resources/XMLParser/2-one-move-no-ranks.pnd
